requirements.txt
streamlit==1.39.0
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
streamlit-drawable-canvas>=0.9.3